requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.13.1,<3.14
drf-spectacular>=0.22.1,<0.23
psycopg2>=2.8.6,<2.9
boto3==1.17.22
botocore==1.20.22
dj-database-url==0.5.0 
django-heroku==0.3.1
django-storages==1.11.1
gunicorn==20.0.4
numpy>=1.22.1,<1.23
pandas>=1.3.1,<1.4
matplotlib
requests==2.25.1
s3transfer==0.3.4
whitenoise==5.2.0 
scikit-learn
category_encoders
# findspark
# pyspark
